+++ src/utils/colorGenerator.ts
@@ -1,4 +1,4 @@
-import { getRotatorValue, setRotatorValue, loadUsedColors } from '../db' // Import necessary functions
+import { getRotatorValue, setRotatorValue, loadUsedColors } from '../db'
 
 // Define the UsedColor interface
 interface UsedColor {
@@ -143,4 +143,11 @@ function hexToHSL(hex: string): UsedColor {
         s: Math.round(s * 100),
         l: Math.round(l * 100),
     }
+}
+
+export function hexToRgba(hex: string, alpha: number): string {
+    const r = parseInt(hex.slice(1, 3), 16)
+    const g = parseInt(hex.slice(3, 5), 16)
+    const b = parseInt(hex.slice(5, 7), 16)
+    return `rgba(${r}, ${g}, ${b}, ${alpha})`
 }